requirements.txt
streamlit
fastapi
uvicorn
pydantic
python-dotenv
langchain_groq
black 
autopep8 
pygments
streamlit-ace
streamlit-code-editor
pygments

twisted
pytest
pytest-asyncio
pytest-cov
httpx
aiohttp
pytest-mock
httpcore
requests
python-multipart
markdown
numpy
pandas
uvicorn

graphviz
plotly
networkx
graphviz
rich